refactor(components): change name module and reorganization of folders
