+++ src/app/epsilon-ui/models/config-logo.model.ts
